requirements.txt
pick==2.2.0
python-decouple==3.8
